# Improve performance of ConfigProvider.fromMap